Do not use physical resolutions (nm)
A lot of cloud-volume/igneous operation does not work well with
float number. We create filenames from bbox and prefer to have
integer numbers, but when using float resolution, the bboxes in
physical dimension are likely to be float number as well. There
are various places in cloud-volume simply floor the float number
into integers, this causes various off-by-n errors with genuinely
float numbers. For example, bounds(mip) * resolution(mip) suppose
to give us the physical bounds, but the value can be off by the
resolution of the mip level. Also the bbox calculated in this way
can be different from the number one get from iterating within
a larger bbox due to rounding (see index_file_paths_for_bbox)
It is less error-prone to stick with the mip level used to generate
the skeletons.